requirements.txt
langchain
langchain-google-genai
langchain-community
python-dotenv
gradio
pandas
tavily-python
Youtube-python